+++ src/uaprs.pas
@@ -560,6 +560,9 @@ begin
       APRSMessageObject.Track.Visible := False;
       APRSMessageObject.Track.LineWidth := 1;
       APRSMessageObject.ModeS := False;
+      APRSMessageObject.RAWMessages := TStringList.Create;
+      APRSMessageObject.RAWMessages.Add(Message[3]);
+      APRSMessageObject.Count := 0;
 
       if (Pos(DataType, WX) > 0) or (Pos(DataType, WXRaw) > 0) or (Pos(DataType, ItemObject) > 0) then
       begin